[refactor] : Moveable과 RandomMover 패키지 이름 변경 및 통합
