+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,13 +1,11 @@
 package racingcar.domain;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatNoException;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.extension.Moveable;
+import racingcar.domain.move.Moveable;
+
+import static org.assertj.core.api.Assertions.*;
 
 @SuppressWarnings("FieldCanBeLocal")
 @DisplayName("Car 클래스")
