+++ src/test/java/racingcar/domain/RacingGameTest.java
@@ -4,7 +4,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.extension.Moveable;
+import racingcar.domain.move.Moveable;
 
 import java.util.List;
